requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusion and transformers
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# Computer vision
opencv-python>=4.8.0
Pillow>=9.5.0

# DINO and CLIP
dinov2>=1.0.0
clip-by-openai>=1.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.65.0
wandb>=0.15.0

# Evaluation metrics
lpips>=0.1.4
pytorch-fid>=0.3.0

# Jupyter and development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
flake8>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0 